Continuing on splitting handlers
