+++ src/ClassicUO.Client/Network/PacketHandlers/ClientVersion.cs
@@ -1,12 +1,6 @@
-using System.Collections.Generic;
 using ClassicUO.Configuration;
 using ClassicUO.Game;
-using ClassicUO.Game.Managers;
-using ClassicUO.Game.Scenes;
-using ClassicUO.Game.UI.Gumps;
-using ClassicUO.Game.UI.Gumps.Login;
 using ClassicUO.IO;
-using ClassicUO.Utility;
 
 namespace ClassicUO.Network.PacketHandlers;
