+++ src/ClassicUO.Client/Network/PacketHandlers/CloseVendorInterface.cs
@@ -1,0 +1,19 @@
+using ClassicUO.Game;
+using ClassicUO.Game.Managers;
+using ClassicUO.Game.UI.Gumps;
+using ClassicUO.IO;
+
+namespace ClassicUO.Network.PacketHandlers;
+
+internal static class CloseVendorInterface
+{
+    public static void Receive(World world, ref StackDataReader p)
+    {
+        if (!world.InGame)
+            return;
+
+        uint serial = p.ReadUInt32BE();
+
+        UIManager.GetGump<ShopGump>(serial)?.Dispose();
+    }
+}
