+++ src/ClassicUO.Client/Network/PacketHandlers/HandlerRegistry.cs
@@ -8,131 +8,129 @@ public delegate void PacketHandler(World world, ref StackDataReader reader);
 
 public static class PacketHandlerRegistry
 {
-    public static IReadOnlyList<(uint Id, PacketHandler Handler)> GetHandlers()
-    {
-        return [
-            (0x1B, EnterWorld.Receive),
-            (0x55, LoginComplete.Receive),
-            (0xBD, ClientVersion.Receive),
-            (0x03, ClientTalk.Receive),
-            (0x0B, Damage.Receive),
-            (0x11, CharacterStatus.Receive),
-            (0x15, FollowR.Receive),
-            (0x16, NewHealthBarUpdate.Receive),
-            (0x17, NewHealthBarUpdate.Receive),
-            (0x1A, UpdateItem.Receive),
-            (0x1C, Talk.Receive),
-            (0x1D, DeleteObject.Receive),
-            (0x20, UpdatePlayer.Receive),
-            (0x21, DenyWalk.Receive),
-            (0x22, ConfirmWalk.Receive),
-            (0x23, DragAnimation.Receive),
-            (0x24, OpenContainer.Receive),
-            (0x25, UpdateContainedItem.Receive),
-            (0x27, DenyMoveItem),
-            (0x28, EndDraggingItem),
-            (0x29, DropItemAccepted),
-            (0x2C, DeathScreen),
-            (0x2D, MobileAttributes),
-            (0x2E, EquipItem),
-            (0x2F, Swing),
-            (0x32, Unknown_0x32),
-            (0x38, Pathfinding),
-            (0x3A, UpdateSkills),
-            (0x3B, CloseVendorInterface),
-            (0x3C, UpdateContainedItems),
-            (0x4E, PersonalLightLevel),
-            (0x4F, LightLevel),
-            (0x54, PlaySoundEffect),
-            (0x56, MapData),
-            (0x5B, SetTime),
-            (0x65, SetWeather),
-            (0x66, BookData),
-            (0x6C, TargetCursor),
-            (0x6D, PlayMusic),
-            (0x6F, SecureTrading),
-            (0x6E, CharacterAnimation),
-            (0x70, GraphicEffect),
-            (0x71, BulletinBoardData),
-            (0x72, Warmode),
-            (0x73, Ping),
-            (0x74, BuyList),
-            (0x77, UpdateCharacter),
-            (0x78, UpdateObject),
-            (0x7C, OpenMenu),
-            (0x88, OpenPaperdoll),
-            (0x89, CorpseEquipment),
-            (0x90, DisplayMap),
-            (0x93, OpenBook),
-            (0x95, DyeData),
-            (0x97, MovePlayer),
-            (0x98, UpdateName),
-            (0x99, MultiPlacement),
-            (0x9A, ASCIIPrompt),
-            (0x9E, SellList),
-            (0xA1, UpdateHitpoints),
-            (0xA2, UpdateMana),
-            (0xA3, UpdateStamina),
-            (0xA5, OpenUrl),
-            (0xA6, TipWindow),
-            (0xAA, AttackCharacter),
-            (0xAB, TextEntryDialog),
-            (0xAF, DisplayDeath),
-            (0xAE, UnicodeTalk),
-            (0xB0, OpenGump),
-            (0xB2, ChatMessage),
-            (0xB7, Help),
-            (0xB8, CharacterProfile),
-            (0xB9, EnableLockedFeatures),
-            (0xBA, DisplayQuestArrow),
-            (0xBB, UltimaMessengerR),
-            (0xBC, Season),
-            (0xBE, AssistVersion),
-            (0xBF, ExtendedCommand),
-            (0xC0, GraphicEffect),
-            (0xC1, DisplayClilocString),
-            (0xC2, UnicodePrompt),
-            (0xC4, Semivisible),
-            (0xC6, InvalidMapEnable),
-            (0xC7, GraphicEffect),
-            (0xC8, ClientViewRange),
-            (0xCA, GetUserServerPingGodClientR),
-            (0xCB, GlobalQueCount),
-            (0xCC, DisplayClilocString),
-            (0xD0, ConfigurationFileR),
-            (0xD1, Logout),
-            (0xD2, UpdateCharacter),
-            (0xD3, UpdateObject),
-            (0xD4, OpenBook),
-            (0xD6, MegaCliLoc.Receive),
-            (0xD7, GenericAOSCommandsR),
-            (0xD8, CustomHouse),
-            (0xDB, CharacterTransferLog),
-            (0xDC, OPLInfo),
-            (0xDD, OpenCompressedGump),
-            (0xDE, UpdateMobileStatus),
-            (0xDF, BuffDebuff),
-            (0xE2, NewCharacterAnimation),
-            (0xE3, KREncryptionResponse),
-            (0xE5, DisplayWaypoint),
-            (0xE6, RemoveWaypoint),
-            (0xF0, KrriosClientSpecial),
-            (0xF1, FreeshardListR),
-            (0xF3, UpdateItemSA),
-            (0xF5, DisplayMap),
-            (0xF6, BoatMoving),
-            (0xF7, PacketList),
-            (EnhancedPacketHandler.EPID, EnhancedPacketHandler.Handle), //For handling custom packets
+    public static IReadOnlyList<(uint Id, PacketHandler Handler)> GetHandlers() =>
+    [
+        (0x1B, EnterWorld.Receive),
+        (0x55, LoginComplete.Receive),
+        (0xBD, ClientVersion.Receive),
+        (0x03, ClientTalk.Receive),
+        (0x0B, Damage.Receive),
+        (0x11, CharacterStatus.Receive),
+        (0x15, FollowR.Receive),
+        (0x16, NewHealthBarUpdate.Receive),
+        (0x17, NewHealthBarUpdate.Receive),
+        (0x1A, UpdateItem.Receive),
+        (0x1C, Talk.Receive),
+        (0x1D, DeleteObject.Receive),
+        (0x20, UpdatePlayer.Receive),
+        (0x21, DenyWalk.Receive),
+        (0x22, ConfirmWalk.Receive),
+        (0x23, DragAnimation.Receive),
+        (0x24, OpenContainer.Receive),
+        (0x25, UpdateContainedItem.Receive),
+        (0x27, DenyMoveItem.Receive),
+        (0x28, EndDraggingItem.Receive),
+        (0x29, DropItemAccepted.Receive),
+        (0x2C, DeathScreen.Receive),
+        (0x2D, MobileAttributes.Receive),
+        (0x2E, EquipItem.Receive),
+        (0x2F, Swing.Receive),
+        (0x32, Unknown032.Receive),
+        (0x38, Pathfinding.Receive),
+        (0x3A, UpdateSkills.Receive),
+        (0x3B, CloseVendorInterface.Receive),
+        (0x3C, UpdateContainedItems.Receive),
+        (0x4E, PersonalLightLevel.Receive),
+        (0x4F, LightLevel.Receive),
+        (0x54, PlaySoundEffect),
+        (0x56, MapData),
+        (0x5B, SetTime),
+        (0x65, SetWeather),
+        (0x66, BookData),
+        (0x6C, TargetCursor),
+        (0x6D, PlayMusic),
+        (0x6F, SecureTrading),
+        (0x6E, CharacterAnimation),
+        (0x70, GraphicEffect),
+        (0x71, BulletinBoardData),
+        (0x72, Warmode),
+        (0x73, Ping),
+        (0x74, BuyList),
+        (0x77, UpdateCharacter),
+        (0x78, UpdateObject),
+        (0x7C, OpenMenu),
+        (0x88, OpenPaperdoll),
+        (0x89, CorpseEquipment),
+        (0x90, DisplayMap),
+        (0x93, OpenBook),
+        (0x95, DyeData),
+        (0x97, MovePlayer),
+        (0x98, UpdateName),
+        (0x99, MultiPlacement),
+        (0x9A, ASCIIPrompt),
+        (0x9E, SellList),
+        (0xA1, UpdateHitpoints),
+        (0xA2, UpdateMana),
+        (0xA3, UpdateStamina),
+        (0xA5, OpenUrl),
+        (0xA6, TipWindow),
+        (0xAA, AttackCharacter),
+        (0xAB, TextEntryDialog),
+        (0xAF, DisplayDeath),
+        (0xAE, UnicodeTalk),
+        (0xB0, OpenGump),
+        (0xB2, ChatMessage),
+        (0xB7, Help),
+        (0xB8, CharacterProfile),
+        (0xB9, EnableLockedFeatures),
+        (0xBA, DisplayQuestArrow),
+        (0xBB, UltimaMessengerR),
+        (0xBC, Season),
+        (0xBE, AssistVersion),
+        (0xBF, ExtendedCommand),
+        (0xC0, GraphicEffect),
+        (0xC1, DisplayClilocString),
+        (0xC2, UnicodePrompt),
+        (0xC4, Semivisible),
+        (0xC6, InvalidMapEnable),
+        (0xC7, GraphicEffect),
+        (0xC8, ClientViewRange),
+        (0xCA, GetUserServerPingGodClientR),
+        (0xCB, GlobalQueCount),
+        (0xCC, DisplayClilocString),
+        (0xD0, ConfigurationFileR),
+        (0xD1, Logout),
+        (0xD2, UpdateCharacter),
+        (0xD3, UpdateObject),
+        (0xD4, OpenBook),
+        (0xD6, MegaCliLoc.Receive),
+        (0xD7, GenericAOSCommandsR),
+        (0xD8, CustomHouse),
+        (0xDB, CharacterTransferLog),
+        (0xDC, OPLInfo),
+        (0xDD, OpenCompressedGump),
+        (0xDE, UpdateMobileStatus),
+        (0xDF, BuffDebuff),
+        (0xE2, NewCharacterAnimation),
+        (0xE3, KREncryptionResponse),
+        (0xE5, DisplayWaypoint),
+        (0xE6, RemoveWaypoint),
+        (0xF0, KrriosClientSpecial),
+        (0xF1, FreeshardListR),
+        (0xF3, UpdateItemSA),
+        (0xF5, DisplayMap),
+        (0xF6, BoatMoving),
+        (0xF7, PacketList),
+        (EnhancedPacketHandler.EPID, EnhancedPacketHandler.Handle), //For handling custom packets
 
-            // login
-            (0xA8, ServerListReceived),
-            (0x8C, ReceiveServerRelay),
-            (0x86, UpdateCharacterList),
-            (0xA9, ReceiveCharacterList),
-            (0x82, ReceiveLoginRejection),
-            (0x85, ReceiveLoginRejection),
-            (0x53, ReceiveLoginRejection),
-            (0xFD, LoginDelay),
-        ];
-    }
+        // login
+        (0xA8, ServerListReceived),
+        (0x8C, ReceiveServerRelay),
+        (0x86, UpdateCharacterList),
+        (0xA9, ReceiveCharacterList),
+        (0x82, ReceiveLoginRejection),
+        (0x85, ReceiveLoginRejection),
+        (0x53, ReceiveLoginRejection),
+        (0xFD, LoginDelay)
+    ];
 }
