+++ src/ClassicUO.Client/Network/PacketHandlers/PacketHandlers.cs
@@ -367,655 +367,6 @@ sealed class PacketHandlers
         }
     }
 
-    private static void DenyMoveItem(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        Item firstItem = world.Items.Get(Client.Game.UO.GameCursor.ItemHold.Serial);
-
-        if (
-            Client.Game.UO.GameCursor.ItemHold.Enabled
-            || Client.Game.UO.GameCursor.ItemHold.Dropped
-                && (firstItem == null || !firstItem.AllowedToDraw)
-        )
-        {
-            if (world.ObjectToRemove == Client.Game.UO.GameCursor.ItemHold.Serial)
-            {
-                world.ObjectToRemove = 0;
-            }
-
-            if (
-                SerialHelper.IsValid(Client.Game.UO.GameCursor.ItemHold.Serial)
-                && Client.Game.UO.GameCursor.ItemHold.Graphic != 0xFFFF
-            )
-            {
-                if (!Client.Game.UO.GameCursor.ItemHold.UpdatedInWorld)
-                {
-                    if (
-                        Client.Game.UO.GameCursor.ItemHold.Layer == Layer.Invalid
-                        && SerialHelper.IsValid(Client.Game.UO.GameCursor.ItemHold.Container)
-                    )
-                    {
-                        // Server should send an UpdateContainedItem after this packet.
-                        Console.WriteLine("=== DENY === ADD TO CONTAINER");
-
-                        AddItemToContainer(
-                            world,
-                            Client.Game.UO.GameCursor.ItemHold.Serial,
-                            Client.Game.UO.GameCursor.ItemHold.Graphic,
-                            Client.Game.UO.GameCursor.ItemHold.TotalAmount,
-                            Client.Game.UO.GameCursor.ItemHold.X,
-                            Client.Game.UO.GameCursor.ItemHold.Y,
-                            Client.Game.UO.GameCursor.ItemHold.Hue,
-                            Client.Game.UO.GameCursor.ItemHold.Container
-                        );
-
-                        UIManager
-                            .GetGump<ContainerGump>(Client.Game.UO.GameCursor.ItemHold.Container)
-                            ?.RequestUpdateContents();
-                    }
-                    else
-                    {
-                        Item item = world.GetOrCreateItem(
-                            Client.Game.UO.GameCursor.ItemHold.Serial
-                        );
-
-                        item.Graphic = Client.Game.UO.GameCursor.ItemHold.Graphic;
-                        item.Hue = Client.Game.UO.GameCursor.ItemHold.Hue;
-                        item.Amount = Client.Game.UO.GameCursor.ItemHold.TotalAmount;
-                        item.Flags = Client.Game.UO.GameCursor.ItemHold.Flags;
-                        item.Layer = Client.Game.UO.GameCursor.ItemHold.Layer;
-                        item.X = Client.Game.UO.GameCursor.ItemHold.X;
-                        item.Y = Client.Game.UO.GameCursor.ItemHold.Y;
-                        item.Z = Client.Game.UO.GameCursor.ItemHold.Z;
-                        item.CheckGraphicChange();
-
-                        Entity container = world.Get(Client.Game.UO.GameCursor.ItemHold.Container);
-
-                        if (container != null)
-                        {
-                            if (SerialHelper.IsMobile(container.Serial))
-                            {
-                                Console.WriteLine("=== DENY === ADD TO PAPERDOLL");
-
-                                world.RemoveItemFromContainer(item);
-                                container.PushToBack(item);
-                                item.Container = container.Serial;
-
-                                UIManager.GetGump<PaperDollGump>(item.Container)?.RequestUpdateContents();
-                                UIManager.GetGump<ModernPaperdoll>(item.Container)?.RequestUpdateContents();
-                            }
-                            else
-                            {
-                                Console.WriteLine("=== DENY === SOMETHING WRONG");
-
-                                world.RemoveItem(item, true);
-                            }
-                        }
-                        else
-                        {
-                            Console.WriteLine("=== DENY === ADD TO TERRAIN");
-
-                            world.RemoveItemFromContainer(item);
-
-                            item.SetInWorldTile(item.X, item.Y, item.Z);
-                        }
-                    }
-                }
-            }
-            else
-            {
-                Log.Error(
-                    $"Wrong data: serial = {Client.Game.UO.GameCursor.ItemHold.Serial:X8}  -  graphic = {Client.Game.UO.GameCursor.ItemHold.Graphic:X4}"
-                );
-            }
-
-            UIManager.GetGump<SplitMenuGump>(Client.Game.UO.GameCursor.ItemHold.Serial)?.Dispose();
-
-            Client.Game.UO.GameCursor.ItemHold.Clear();
-        }
-        else
-        {
-            Log.Warn("There was a problem with ItemHold object. It was cleared before :|");
-        }
-
-        //var result = World.Items.Get(ItemHold.Serial);
-
-        //if (result != null && !result.IsDestroyed)
-        //    result.AllowedToDraw = true;
-
-        byte code = p.ReadUInt8();
-
-        if (code < 5)
-        {
-            world.MessageManager.HandleMessage(
-                null,
-                ServerErrorMessages.GetError(p[0], code),
-                string.Empty,
-                0x03b2,
-                MessageType.System,
-                3,
-                TextType.SYSTEM
-            );
-        }
-    }
-
-    private static void EndDraggingItem(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        Client.Game.UO.GameCursor.ItemHold.Enabled = false;
-        Client.Game.UO.GameCursor.ItemHold.Dropped = false;
-    }
-
-    private static void DropItemAccepted(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        Client.Game.UO.GameCursor.ItemHold.Enabled = false;
-        Client.Game.UO.GameCursor.ItemHold.Dropped = false;
-
-        Console.WriteLine("PACKET - ITEM DROP OK!");
-    }
-
-    private static void DeathScreen(World world, ref StackDataReader p)
-    {
-        // todo
-        byte action = p.ReadUInt8();
-
-        if (action != 1)
-        {
-            world.Weather.Reset();
-
-            Client.Game.Audio.PlayMusic(Client.Game.Audio.DeathMusicIndex, true);
-
-            if (ProfileManager.CurrentProfile.EnableDeathScreen)
-            {
-                world.Player.DeathScreenTimer = Time.Ticks + Constants.DEATH_SCREEN_TIMER;
-            }
-
-            GameActions.RequestWarMode(world.Player, false);
-            world.WMapManager._corpse = new WMapEntity(world.Player.Serial)
-            {
-                X = world.Player.X,
-                Y = world.Player.Y,
-                HP = 0,
-                Map = world.Map.Index,
-                LastUpdate = Time.Ticks + (1000 * 60 * 5),
-                IsGuild = false,
-                Name = $"Your Corpse"
-            };
-
-            EventSink.InvokeOnPlayerDeath(world.Player, world.Player.Serial);
-        }
-    }
-
-    private static void MobileAttributes(World world, ref StackDataReader p)
-    {
-        uint serial = p.ReadUInt32BE();
-
-        Entity entity = world.Get(serial);
-
-        if (entity == null)
-        {
-            return;
-        }
-
-        ushort oldHits = entity.Hits;
-        entity.HitsMax = p.ReadUInt16BE();
-        entity.Hits = p.ReadUInt16BE();
-
-        if (entity.HitsRequest == HitsRequestStatus.Pending)
-        {
-            entity.HitsRequest = HitsRequestStatus.Received;
-        }
-
-        if (SerialHelper.IsMobile(serial))
-        {
-            var mobile = entity as Mobile;
-
-            if (mobile == null)
-            {
-                return;
-            }
-
-            mobile.ManaMax = p.ReadUInt16BE();
-            mobile.Mana = p.ReadUInt16BE();
-            mobile.StaminaMax = p.ReadUInt16BE();
-            mobile.Stamina = p.ReadUInt16BE();
-
-            if (mobile == world.Player)
-            {
-                TitleBarStatsManager.UpdateTitleBar();
-            }
-
-            // Check for bandage healing
-            if (oldHits != mobile.Hits)
-            {
-                BandageManager.Instance.OnMobileHpChanged(mobile, oldHits, mobile.Hits);
-            }
-        }
-    }
-
-    private static void EquipItem(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        uint serial = p.ReadUInt32BE();
-
-        Item item = world.GetOrCreateItem(serial);
-
-        if (item.Graphic != 0 && item.Layer != Layer.Backpack)
-        {
-            //ClearContainerAndRemoveItems(item);
-            world.RemoveItemFromContainer(item);
-        }
-
-        if (SerialHelper.IsValid(item.Container))
-        {
-            UIManager.GetGump<ContainerGump>(item.Container)?.RequestUpdateContents();
-
-            UIManager.GetGump<PaperDollGump>(item.Container)?.RequestUpdateContents();
-            UIManager.GetGump<ModernPaperdoll>(item.Container)?.RequestUpdateContents();
-        }
-
-        item.Graphic = (ushort)(p.ReadUInt16BE() + p.ReadInt8());
-        item.Layer = (Layer)p.ReadUInt8();
-        item.Container = p.ReadUInt32BE();
-        item.FixHue(p.ReadUInt16BE());
-        item.Amount = 1;
-
-        Entity entity = world.Get(item.Container);
-
-        entity?.PushToBack(item);
-
-        if (item.Layer == Layer.Mount && entity is Mobile mob)
-        {
-            mob.Mount = item;
-        }
-
-        if (item.Layer >= Layer.ShopBuyRestock && item.Layer <= Layer.ShopSell)
-        {
-            //item.Clear();
-        }
-        else if (SerialHelper.IsValid(item.Container) && item.Layer < Layer.Mount)
-        {
-            UIManager.GetGump<PaperDollGump>(item.Container)?.RequestUpdateContents();
-            UIManager.GetGump<ModernPaperdoll>(item.Container)?.RequestUpdateContents();
-        }
-
-        if (
-            entity == world.Player
-            && (item.Layer == Layer.OneHanded || item.Layer == Layer.TwoHanded)
-        )
-        {
-            world.Player?.UpdateAbilities();
-        }
-
-        //if (ItemHold.Serial == item.Serial)
-        //{
-        //    Console.WriteLine("PACKET - ITEM EQUIP");
-        //    ItemHold.Enabled = false;
-        //    ItemHold.Dropped = true;
-        //}
-    }
-
-    private static void Swing(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        p.Skip(1);
-
-        uint attackers = p.ReadUInt32BE();
-
-        if (attackers != world.Player)
-        {
-            return;
-        }
-
-        uint defenders = p.ReadUInt32BE();
-
-        const int TIME_TURN_TO_LASTTARGET = 2000;
-
-        if (
-            world.TargetManager.LastAttack == defenders
-            && world.Player.InWarMode
-            && world.Player.Walker.LastStepRequestTime + TIME_TURN_TO_LASTTARGET < Time.Ticks
-            && world.Player.Steps.Count == 0
-        )
-        {
-            Mobile enemy = world.Mobiles.Get(defenders);
-
-            if (enemy != null)
-            {
-                Direction pdir = DirectionHelper.GetDirectionAB(
-                    world.Player.X,
-                    world.Player.Y,
-                    enemy.X,
-                    enemy.Y
-                );
-
-                int x = world.Player.X;
-                int y = world.Player.Y;
-                sbyte z = world.Player.Z;
-
-                if (
-                    world.Player.Pathfinder.CanWalk(ref pdir, ref x, ref y, ref z)
-                    && world.Player.Direction != pdir
-                )
-                {
-                    world.Player.Walk(pdir, false);
-                }
-            }
-        }
-    }
-
-    private static void Unknown_0x32(World world, ref StackDataReader p) { }
-
-    private static void UpdateSkills(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        byte type = p.ReadUInt8();
-        bool haveCap = type != 0u && type <= 0x03 || type == 0xDF;
-        bool isSingleUpdate = type == 0xFF || type == 0xDF;
-
-        if (type == 0xFE)
-        {
-            int count = p.ReadUInt16BE();
-
-            Client.Game.UO.FileManager.Skills.Skills.Clear();
-            Client.Game.UO.FileManager.Skills.SortedSkills.Clear();
-
-            for (int i = 0; i < count; i++)
-            {
-                bool haveButton = p.ReadBool();
-                int nameLength = p.ReadUInt8();
-
-                Client.Game.UO.FileManager.Skills.Skills.Add(
-                    new SkillEntry(i, p.ReadASCII(nameLength), haveButton)
-                );
-            }
-
-            Client.Game.UO.FileManager.Skills.SortedSkills.AddRange(Client.Game.UO.FileManager.Skills.Skills);
-
-            Client.Game.UO.FileManager.Skills.SortedSkills.Sort(
-                (a, b) => string.Compare(a.Name, b.Name, StringComparison.InvariantCulture)
-            );
-        }
-        else
-        {
-            StandardSkillsGump standard = null;
-            SkillGumpAdvanced advanced = null;
-
-            if (ProfileManager.CurrentProfile.StandardSkillsGump)
-            {
-                standard = UIManager.GetGump<StandardSkillsGump>();
-            }
-            else
-            {
-                advanced = UIManager.GetGump<SkillGumpAdvanced>();
-            }
-
-            if (!isSingleUpdate && (type == 1 || type == 3 || world.SkillsRequested))
-            {
-                world.SkillsRequested = false;
-
-                // TODO: make a base class for this gump
-                if (ProfileManager.CurrentProfile.StandardSkillsGump)
-                {
-                    if (standard == null)
-                    {
-                        UIManager.Add(standard = new StandardSkillsGump(world) { X = 100, Y = 100 });
-                    }
-                }
-                else
-                {
-                    if (advanced == null)
-                    {
-                        UIManager.Add(advanced = new SkillGumpAdvanced(world) { X = 100, Y = 100 });
-                    }
-                }
-            }
-
-            while (p.Position < p.Length)
-            {
-                ushort id = p.ReadUInt16BE();
-
-                if (p.Position >= p.Length)
-                {
-                    break;
-                }
-
-                if (id == 0 && type == 0)
-                {
-                    break;
-                }
-
-                if (type == 0 || type == 0x02)
-                {
-                    id--;
-                }
-
-                ushort realVal = p.ReadUInt16BE();
-                ushort baseVal = p.ReadUInt16BE();
-                var locked = (Lock)p.ReadUInt8();
-                ushort cap = 1000;
-
-                if (haveCap)
-                {
-                    cap = p.ReadUInt16BE();
-                }
-
-                if (id < world.Player.Skills.Length)
-                {
-                    Skill skill = world.Player.Skills[id];
-
-                        if (skill != null)
-                        {
-                            if (isSingleUpdate)
-                            {
-                                float change = realVal / 10.0f - skill.Value;
-                                int deltaThreshold = ProfileManager.CurrentProfile?.ShowSkillsChangedDeltaValue ?? 0;
-
-                                if (
-                                    change != 0.0f
-                                    && !float.IsNaN(change)
-                                    && ProfileManager.CurrentProfile != null
-                                    && ProfileManager.CurrentProfile.ShowSkillsChangedMessage
-                                    && (
-                                        deltaThreshold <= 0
-                                        || skill.ValueFixed / deltaThreshold != realVal / deltaThreshold
-                                    )
-                                )
-                                {
-                                    GameActions.Print(
-                                        world,
-                                        string.Format(
-                                            ResGeneral.YourSkillIn0Has1By2ItIsNow3,
-                                            skill.Name,
-                                            change < 0
-                                                ? ResGeneral.Decreased
-                                                : ResGeneral.Increased,
-                                            Math.Abs(change),
-                                            skill.Value + change
-                                        ),
-                                        0x58,
-                                        MessageType.System,
-                                        3,
-                                        false
-                                    );
-                                }
-                            }
-
-                        ushort lastBase = skill.BaseFixed;
-                        ushort lastValue = skill.ValueFixed;
-                        ushort lastCap = skill.CapFixed;
-
-                        skill.BaseFixed = baseVal;
-                        skill.ValueFixed = realVal;
-                        skill.CapFixed = cap;
-                        skill.Lock = locked;
-
-                        if (isSingleUpdate)
-                        {
-                            if (lastBase != skill.BaseFixed)
-                                Skill.InvokeSkillBaseChanged(id);
-                            if (lastValue != skill.ValueFixed)
-                                Skill.InvokeSkillValueChanged(id);
-                            if (lastCap != skill.CapFixed)
-                                Skill.InvokeSkillCapChanged(id);
-                        }
-
-                        standard?.Update(id);
-                        advanced?.ForceUpdate();
-                    }
-                }
-
-                if (isSingleUpdate)
-                {
-                    break;
-                }
-            }
-        }
-    }
-
-    private static void Pathfinding(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        ushort x = p.ReadUInt16BE();
-        ushort y = p.ReadUInt16BE();
-        ushort z = p.ReadUInt16BE();
-
-        world.Player.Pathfinder.WalkTo(x, y, z, 0);
-    }
-
-    private static void UpdateContainedItems(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        ushort count = p.ReadUInt16BE();
-
-        for (int i = 0; i < count; i++)
-        {
-            uint serial = p.ReadUInt32BE();
-            ushort graphic = (ushort)(p.ReadUInt16BE() + p.ReadUInt8());
-            ushort amount = Math.Max(p.ReadUInt16BE(), (ushort)1);
-            ushort x = p.ReadUInt16BE();
-            ushort y = p.ReadUInt16BE();
-
-            if (Client.Game.UO.Version >= Utility.ClientVersion.CV_6017)
-            {
-                p.Skip(1);
-            }
-
-            uint containerSerial = p.ReadUInt32BE();
-            ushort hue = p.ReadUInt16BE();
-
-            if (i == 0)
-            {
-                Entity container = world.Get(containerSerial);
-
-                if (container != null)
-                {
-                    ClearContainerAndRemoveItems(world, container, container.Graphic == 0x2006);
-                }
-            }
-
-            AddItemToContainer(world, serial, graphic, amount, x, y, hue, containerSerial);
-        }
-    }
-
-    private static void CloseVendorInterface(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        uint serial = p.ReadUInt32BE();
-
-        UIManager.GetGump<ShopGump>(serial)?.Dispose();
-    }
-
-    private static void PersonalLightLevel(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        if (world.Player == p.ReadUInt32BE())
-        {
-            byte level = p.ReadUInt8();
-
-            if (level > 0x1E)
-            {
-                level = 0x1E;
-            }
-
-            world.Light.RealPersonal = level;
-
-            if (!ProfileManager.CurrentProfile.UseCustomLightLevel)
-            {
-                world.Light.Personal = level;
-            }
-        }
-    }
-
-    private static void LightLevel(World world, ref StackDataReader p)
-    {
-        if (!world.InGame)
-        {
-            return;
-        }
-
-        byte level = p.ReadUInt8();
-
-        if (level > 0x1E)
-        {
-            level = 0x1E;
-        }
-
-        world.Light.RealOverall = level;
-
-        if (
-            !ProfileManager.CurrentProfile.UseCustomLightLevel
-            || ProfileManager.CurrentProfile.LightLevelType == 1
-        )
-        {
-            world.Light.Overall =
-                ProfileManager.CurrentProfile.LightLevelType == 1
-                    ? Math.Min(level, ProfileManager.CurrentProfile.LightLevel)
-                    : level;
-        }
-    }
-
     private static void PlaySoundEffect(World world, ref StackDataReader p)
     {
         if (world.Player == null)
