+++ src/ClassicUO.Client/Network/PacketHandlers/Unknown032.cs
@@ -1,0 +1,11 @@
+using ClassicUO.Game;
+using ClassicUO.IO;
+
+namespace ClassicUO.Network.PacketHandlers;
+
+internal static class Unknown032
+{
+    public static void Receive(World world, ref StackDataReader p)
+    {
+    }
+}
